+++ components/Comment/lib/types.ts
@@ -10,6 +10,10 @@ import {
 import { parseVote, Vote } from "~/config/types/vote";
 import { formatDateStandard, timeSince } from "~/config/utils/dates";
 import { isEmpty } from "~/config/utils/nullchecks";
+import {
+  parseAnchor,
+  SerializedAnchorPosition,
+} from "../modules/annotation/lib/types";
 
 export enum COMMENT_TYPES {
   DISCUSSION = "GENERIC_COMMENT",
@@ -32,51 +36,6 @@ export enum COMMENT_CONTEXTS {
   ANNOTATION = "ANNOTATION",
   FEED = "FEED",
 }
-
-export type Annotation = {
-  threadId: string | "new-annotation";
-  serialized: SerializedAnchorPosition;
-  anchorCoordinates: Array<{
-    x: number;
-    y: number;
-    width: number;
-    height: number;
-  }>;
-  threadCoordinates: {
-    x: number;
-    y: number;
-  };
-  xrange: any;
-};
-
-type SerializedAnchorPosition = {
-  startContainerPath: string;
-  startOffset: number;
-  endContainerPath: string;
-  endOffset: number;
-  collapsed: boolean;
-  textContent: string;
-  page?: number;
-  type: "text";
-};
-
-// export type AnnotationAnchorPosition = {
-//   threadId: ID | "new-thread";
-//   serialized: SerializedAnchorPosition;
-//   anchorCoordinates: Array<{
-//     x: number;
-//     y: number;
-//     width: number;
-//     height: number;
-//   }>;
-//   xrange: any;
-// };
-
-// export type AnnotationThreadPosition = {
-//   threadId: ID | "new-thread";
-//   x: number;
-//   y: number;
-// }
 
 export type CommentThread = {
   id: ID;
@@ -104,65 +63,6 @@ export type Comment = {
   isAcceptedAnswer: boolean;
   review?: Review;
   thread: CommentThread;
-};
-
-// export type UnrenderedAnnotation = {
-//   comment?: Comment;
-//   xrange: any | null;
-//   isNew?: boolean;
-// };
-
-// export type RenderedAnnotation = {
-//   comment?: Comment;
-//   isNew?: boolean;
-//   xrange: any;
-//   anchorCoordinates: Array<{
-//     x: number;
-//     y: number;
-//     width: number;
-//     height: number;
-//   }>;
-//   commentCoordinates: { x: number; y: number };
-// };
-
-export const createAnnotation = ({
-  serializedAnchorPosition,
-  xrange,
-  relativeEl,
-  threadId,
-}: {
-  xrange: any;
-  relativeEl?: any;
-  threadId?: string | "new-annotation";
-  serializedAnchorPosition?: SerializedAnchorPosition;
-}): Annotation => {
-  const highlightCoords = xrange.getCoordinates({
-    relativeEl: relativeEl,
-  });
-
-  return {
-    threadId: threadId || "new-annotation",
-    serialized: xrange.serialize() || serializedAnchorPosition,
-    anchorCoordinates: highlightCoords,
-    threadCoordinates: {
-      x: 0,
-      y: highlightCoords[0].y, // Initial position on first render before adjustment
-    },
-    xrange,
-  };
-};
-
-export const parseAnchor = (raw: any): SerializedAnchorPosition => {
-  return {
-    type: raw.type,
-    startContainerPath: raw.position?.startContainerPath,
-    startOffset: raw.position?.startOffset,
-    endContainerPath: raw.position?.endContainerPath,
-    endOffset: raw.position?.endOffset,
-    collapsed: raw.position?.collapsed,
-    textContent: raw.position?.textContent,
-    page: raw.position?.page || null,
-  };
 };
 
 export const parseThread = (raw: any): CommentThread => {
